style: 모호한 이름 직관적이게 수정
