+++ src/components/cart/orderSummarySection/OrderSummarySection.tsx
@@ -1,36 +1,7 @@
 import styled from 'styled-components';
 import { getCommaAddedNumber } from '../../../utils/number';
-import { selector, useRecoilValue } from 'recoil';
-import {
-  cartItemsState,
-  selectedCartIdListState,
-} from '../../../atoms/cartAtom';
-
-const priceSummaryState = selector({
-  key: 'priceSummaryState',
-  get: ({ get }) => {
-    const selectedCartItems = get(selectedCartIdListState);
-    const cartItems = get(cartItemsState);
-
-    const totalProductPrice = selectedCartItems.reduce(
-      (acc, selectedCartItemId) => {
-        const product = cartItems.find(
-          (cartProduct) => cartProduct.id === selectedCartItemId
-        );
-
-        return (acc +=
-          Number(product?.quantity) * Number(product?.product.price));
-      },
-      0
-    );
-
-    const deliveryPrice = 3000;
-
-    const totalPrice = totalProductPrice + deliveryPrice;
-
-    return { totalProductPrice, deliveryPrice, totalPrice };
-  },
-});
+import { useRecoilValue } from 'recoil';
+import { priceSummaryState } from '../../../recoil/selectors/priceSummary';
 
 export const OrderSummarySection = () => {
   const { totalProductPrice, deliveryPrice, totalPrice } =
