+++ src/pages/Main.tsx
@@ -1,26 +1,22 @@
-import { ProductCardGrid } from '../components/main/productCardList/ProductCardList';
+import { ProductCardGrid } from '../components/main/productCardGrid/ProductCardGrid';
 import { Layout } from '../layout';
 import { useEffect, useState } from 'react';
 import { Product } from '../types/Product';
 
 function Main() {
-  const [mockData, setMockData] = useState<Product[] | undefined>();
+  const [products, setProducts] = useState<Product[] | undefined>();
 
   useEffect(() => {
     fetch('/products')
       .then((res) => res.json())
       .then((data) => {
-        setMockData(data);
+        setProducts(data);
       });
-
-    fetch('/cart-items')
-      .then((res) => res.json())
-      .then((data) => {});
   }, []);
 
   return (
     <Layout>
-      {mockData ? <ProductCardGrid products={mockData} /> : <>loading...</>}
+      {products ? <ProductCardGrid products={products} /> : <>loading...</>}
     </Layout>
   );
 }
