+++ src/recoil/selectors/priceSummary.ts
@@ -1,0 +1,28 @@
+import { selector } from 'recoil';
+import { cartItemsState, selectedCartIdListState } from '../atoms/cartAtom';
+
+export const priceSummaryState = selector({
+  key: 'priceSummaryState',
+  get: ({ get }) => {
+    const selectedCartItems = get(selectedCartIdListState);
+    const cartItems = get(cartItemsState);
+
+    const totalProductPrice = selectedCartItems.reduce(
+      (acc, selectedCartItemId) => {
+        const product = cartItems.find(
+          (cartProduct) => cartProduct.id === selectedCartItemId
+        );
+
+        return (acc +=
+          Number(product?.quantity) * Number(product?.product.price));
+      },
+      0
+    );
+
+    const deliveryPrice = 3000;
+
+    const totalPrice = totalProductPrice + deliveryPrice;
+
+    return { totalProductPrice, deliveryPrice, totalPrice };
+  },
+});
